pyproject.toml
[project]
name = "opennet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyhamcrest>=2.1.0",
    "pytest>=8.4.1",
    "pyyaml>=6.0.2",
    "selenium>=4.33.0",
]
